pyproject.toml
[project]
name = "smcs-stock-prediction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "en-core-web-sm",
    "huggingface-hub>=1.5.0",
    "pandas>=3.0.1",
    "polars>=1.38.1",
    "pyarrow>=23.0.1",
    "pytextrank>=3.3.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "spacy>=3.8.11",
    "trafilatura>=2.0.0",
]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }
